fix compilation under clang 3.1 (fixes #5)
GCC, bizarrely, does not seem to compile methods in templates that go unused.
The getError method of the FloatingAverage<T> template class has an obvious
coding mistake, but compiling it with GCC never revealed this because getError
is never called in the library. Clang 3.1, however, seems to be somewhat more
conservative about this; it flags an error anyway. Because this method is
unused, I just removed it.